ze/pyproject.toml
[project]
name = "zeconfig"
version = "0.1.4"
description = "Configuration Manager Package"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "rich>=14.2.0",
    "tomli>=2.2.1",
    "typer>=0.20.0",
]

[project.scripts]
ze = "ze.src.cli:main"

[dependency-groups]
dev = [
    "coverage==7.6.12",
    "iniconfig==2.0.0",
    "packaging==24.2",
    "pluggy==1.5.0",
    "psutil==6.1.1",
    "pytest==8.3.4",
    "pytest-cov==6.0.0",
    "ruff==0.9.7",
    "taskipy==1.14.1",
]
docs = [
    "mkdocs>=1.6.1",
    "mkdocs-material>=9.6.7",
    "mkdocstrings[python]>=0.29.0",
]

[tool.setuptools]
packages = ["ze", "ze.src"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
extend-exclude = []

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
pre_test = 'task lint'
test = 'pytest -s -x --cov=ze/src -vv'
post_test = 'coverage html'
lint = 'ruff check'
pre_format = 'ruff check --fix'
format = 'ruff format'
